Revert "Abstract out osu specific stuff from UpdateManager"
This reverts commit 5e38187924f3ca6be80a0e62cccb9ce9ac481656.
