+++ osu.Game/Updater/SimpleUpdateManager.cs
@@ -1,8 +1,13 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+#nullable disable
+
+using System;
 using System.Linq;
+using System.Runtime.InteropServices;
 using System.Threading.Tasks;
+using osu.Framework;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Platform;
@@ -15,28 +20,24 @@ namespace osu.Game.Updater
     /// An update manager that shows notifications if a newer release is detected.
     /// Installation is left up to the user.
     /// </summary>
-    /// <remarks>
-    /// Implementers are expected to set <see cref="Version"/> to something that matches their release versioning/tagging system.
-    /// </remarks>
-    public abstract partial class SimpleUpdateManager : UpdateManager
+    public partial class SimpleUpdateManager : UpdateManager
     {
-        protected string Version { get; set; } = "0.0.0";
-
-        protected abstract string GitHubUrl { get; }
-
-        // For lack of a better term
-        protected abstract string UpdatedComponentName { get; }
-
-        protected virtual string DownloadMessage { get; } = "Click here to download the new version, which can be installed over the top of your existing installation";
+        private string version;
 
         [Resolved]
-        private GameHost host { get; set; } = null!;
+        private GameHost host { get; set; }
+
+        [BackgroundDependencyLoader]
+        private void load(OsuGameBase game)
+        {
+            version = game.Version;
+        }
 
         protected override async Task<bool> PerformUpdateCheck()
         {
             try
             {
-                var releases = new OsuJsonWebRequest<GitHubRelease>(GitHubUrl);
+                var releases = new OsuJsonWebRequest<GitHubRelease>("https://api.github.com/repos/ppy/osu/releases/latest");
 
                 await releases.PerformAsync().ConfigureAwait(false);
 
@@ -44,19 +45,19 @@ namespace osu.Game.Updater
 
                 // avoid any discrepancies due to build suffixes for now.
                 // eventually we will want to support release streams and consider these.
-                Version = Version.Split('-').First();
+                version = version.Split('-').First();
                 string latestTagName = latest.TagName.Split('-').First();
 
-                if (latestTagName != Version)
+                if (latestTagName != version)
                 {
                     Notifications.Post(new SimpleNotification
                     {
-                        Text = $"A newer release of {UpdatedComponentName} has been found ({Version} → {latestTagName}).\n\n"
-                               + DownloadMessage,
+                        Text = $"A newer release of osu! has been found ({version} → {latestTagName}).\n\n"
+                               + "Click here to download the new version, which can be installed over the top of your existing installation",
                         Icon = FontAwesome.Solid.Download,
                         Activated = () =>
                         {
-                            UpdateActionOnClick(latest);
+                            host.OpenUrlExternally(getBestUrl(latest));
                             return true;
                         }
                     });
@@ -73,8 +74,37 @@ namespace osu.Game.Updater
             return false;
         }
 
-        protected virtual void UpdateActionOnClick(GitHubRelease release) => host.OpenUrlExternally(GetBestURL(release));
+        private string getBestUrl(GitHubRelease release)
+        {
+            GitHubAsset bestAsset = null;
 
-        protected virtual string GetBestURL(GitHubRelease release) => release.HtmlUrl;
+            switch (RuntimeInfo.OS)
+            {
+                case RuntimeInfo.Platform.Windows:
+                    bestAsset = release.Assets?.Find(f => f.Name.EndsWith(".exe", StringComparison.Ordinal));
+                    break;
+
+                case RuntimeInfo.Platform.macOS:
+                    string arch = RuntimeInformation.OSArchitecture == Architecture.Arm64 ? "Apple.Silicon" : "Intel";
+                    bestAsset = release.Assets?.Find(f => f.Name.EndsWith($".app.{arch}.zip", StringComparison.Ordinal));
+                    break;
+
+                case RuntimeInfo.Platform.Linux:
+                    bestAsset = release.Assets?.Find(f => f.Name.EndsWith(".AppImage", StringComparison.Ordinal));
+                    break;
+
+                case RuntimeInfo.Platform.iOS:
+                    // iOS releases are available via testflight. this link seems to work well enough for now.
+                    // see https://stackoverflow.com/a/32960501
+                    return "itms-beta://beta.itunes.apple.com/v1/app/1447765923";
+
+                case RuntimeInfo.Platform.Android:
+                    // on our testing device this causes the download to magically disappear.
+                    //bestAsset = release.Assets?.Find(f => f.Name.EndsWith(".apk"));
+                    break;
+            }
+
+            return bestAsset?.BrowserDownloadUrl ?? release.HtmlUrl;
+        }
     }
 }
